Init commits
need to following :
This test is in 4 parts. You will need to complete Part 1 and Part 2 in order before moving onto Part 3 and Part 4. Part 4 is optional, but please note that a good submission will give you the best opportunity to stand out from the rest of the applicants, and we would encourage you to attempt it.
This test will require you to have access to a Windows PC with the following development tools:
•	Visual Studio Community 2017 or better:
https://visualstudio.microsoft.com/vs/community/
•	Windows SDK with DX11 or newer
You should have a zip file containing the source code for a simple Asteroids game. This code will form the basis of the test.
As a rough guide, Part 1 should take no longer than 30 minutes, Part 2 should take no longer than 1 hour, and Part 3 should take no longer than 2 hours. The length of time spent on Part 4 will depend on your choice of change.
You should spend no longer than 2 working days in total on this test. Please make a note of the time you spend on each part of the test and include these times when sending in your submission – we will take into account the amount of time you have been able to spend on the test when evaluating your submission.
Part 1 – Compiling
The source code in the zip file will not compile. Find and fix the compile errors.
There is a compiler warning about sprintf in LevelStart::OnRender. Modify the function to use sprintf_s and get rid of the warning.
Part 2 – Bug Fixing
There are two deliberate bugs in the game that was cause a crash. Find and fix the crashes.
If you are unable to fix the bugs, please write down as much detail as you have been able to gather about the possible causes of the bugs, and what changes you think will fix the bugs.
(HINT: One of the bugs is not just a simple crash. You will be expected to make sure that your fix has not introduced any additional bugs.)
Part 3 – Small Improvements
Currently only one fired bullet can be active at a time. Change the code so that multiple bullets can be active simultaneously.